Intel QAT Z.I.A. provider for ZFS only
This provider takes the original Intel QAT code found in ZFS
and moves it into a Z.I.A. provider. Because the original
code uses symbols from ZFS, this provider should be treated
as only usable by ZFS.

The original QAT code in ZFS did not use the new operation
chaining capability that was added to the QAT API in March
2019. This provider does not update the code to use it as
using chaining would cause operations to be performed out of
sync with ZFS.

The Intel QAT does not maintain memory locally and instead
always reads from in-memory buffers and write results to
in-memory buffers. This results in a lot of memory traffic.
Additionally, because the DPUSM (intentionally) does not have
an interface for passing raw pointers from the caller to the
accelerator, this provider will allocate another in-memory
buffer to copy each offloaded ZFS buffer into, doubling
memory usage.

The vdev_raidz, vdev_file, and vdev_disk code was copied
from the example software provider.

Encryption is not supported by the DPUSM. The original
code was moved along with the rest of the QAT code, but
is now not used by ZFS.

Signed-off-by: Jason Lee <jasonlee@lanl.gov>