get_bdev Function
Adds a function that returns the
correct block device handle from
the vdev_disk_t struct defined
in vdev_disk.c.

The purpose of this addition is
to use this function to get the
block device handle to pass to
zia_disk_open from zia_open_vdevs,
which currently just casts a
vdev_disk_t object to a block device *,
leading to failure if the operation
does not fallback.

Signed-off-by: Luke Beirne <lbeirne@lanl.gov>